Rename namespace Serialization to ApiSerialization
