+++ Adyen.Test/SerializerTest.cs
@@ -21,7 +21,7 @@
 //  */
 #endregion
 
-using Adyen.Serialization;
+using Adyen.ApiSerialization;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Adyen.Model.Nexo;
 using System.Text;
