+++ Adyen/ApiSerialization/IMessagePayloadSerializer.cs
@@ -21,7 +21,7 @@
 //  */
 #endregion
 
-namespace Adyen.Serialization
+namespace Adyen.ApiSerialization
 {
     internal interface IMessagePayloadSerializer<out T> where T : IMessagePayload
     {
