+++ Adyen/ApiSerialization/JSonConvertDeserializerWrapper.cs
@@ -20,32 +20,29 @@
 //  * See the LICENSE file for more info.
 //  */
 #endregion
+
+using Adyen.Exceptions;
 using Newtonsoft.Json;
-using System;
+using Newtonsoft.Json.Serialization;
 
-namespace Adyen.Serialization.Converter
+namespace Adyen.ApiSerialization
 {
-    internal class JsonBase64Converter : JsonConverter
+    internal class JSonConvertDeserializerWrapper<T>
     {
-        public override bool CanConvert(Type objectType)
+        internal static T DeserializeObject(string objectToDeserialize)
         {
-            return true;
-        }
+            return JsonConvert.DeserializeObject<T>(objectToDeserialize, new JsonSerializerSettings
+            {
+                MissingMemberHandling = MissingMemberHandling.Error,
+                Error = delegate (object sender, ErrorEventArgs args)
+                {
+                    var exceptionMessage = string.Format(ExceptionMessages.ExceptionDuringDeserialization,
+                        objectToDeserialize,
+                        args.ErrorContext.Error.Message);
 
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            return System.Text.Encoding.UTF8.GetString((Convert.FromBase64String((string)reader.Value)));
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            var json = JsonConvert.SerializeObject(value, Formatting.None,
-                            new JsonSerializerSettings
-                            {
-                                NullValueHandling = NullValueHandling.Ignore
-                            });
-
-            writer.WriteValue(Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(json)));
+                    throw new DeserializationException(exceptionMessage, args.ErrorContext.Error);
+                }
+            });
         }
     }
 }
