+++ Adyen/ApiSerialization/MessagePayloadSerializer.cs
@@ -21,29 +21,14 @@
 //  */
 #endregion
 
-using System;
-using Adyen.Model.Nexo;
-
-namespace Adyen.Serialization
+namespace Adyen.ApiSerialization
 {
-    internal class MessagePayloadSerializerFactory
+    internal class MessagePayloadSerializer<T> : IMessagePayloadSerializer<T> where T : IMessagePayload
     {
-        internal IMessagePayloadSerializer<IMessagePayload> CreateSerializer(string messageCategory, string messageType)
+        public IMessagePayload Deserialize(string messagePayloadJson)
         {
-            var messagePayoadFullName = CreateMessagePayloadFullName(messageCategory, messageType);
-            var messagePayloadSerializer = TypeHelper.CreateGenericTypeFromStringFullNamespace(typeof(MessagePayloadSerializer<>), messagePayoadFullName);
-
-            return (IMessagePayloadSerializer<IMessagePayload>)Activator.CreateInstance(messagePayloadSerializer);
+            return Converter.JSonConvertDeserializerWrapper<T>.DeserializeObject(messagePayloadJson);
         }
 
-        private string CreateMessagePayloadFullName(string messageCategory, string messageType)
-        {
-            var nameSpaceSeparator = ".";
-
-            var messagePayloadName = messageCategory.ToString() + messageType;
-            var nexoDomainNameSpace = typeof(PaymentRequest).Namespace;
-
-            return string.Concat(nexoDomainNameSpace, nameSpaceSeparator, messagePayloadName);
-        }
     }
 }
